+++ app/src/main/java/com/fieldbook/tracker/traits/formats/DiseaseRatingFormat.kt
@@ -1,7 +1,11 @@
 package com.fieldbook.tracker.traits.formats
 
 import com.fieldbook.tracker.R
+import com.fieldbook.tracker.traits.formats.feature.ChartableData
+import com.fieldbook.tracker.traits.formats.feature.DisplayValue
+import com.fieldbook.tracker.traits.formats.feature.Scannable
 import com.fieldbook.tracker.traits.formats.parameters.DetailsParameter
+import com.fieldbook.tracker.traits.formats.parameters.AttachMediaParameter
 import com.fieldbook.tracker.traits.formats.parameters.NameParameter
 import com.fieldbook.tracker.traits.formats.parameters.RepeatedMeasureParameter
 import com.fieldbook.tracker.traits.formats.parameters.ResourceFileParameter
@@ -19,5 +23,6 @@ class DiseaseRatingFormat : TraitFormat(
     DetailsParameter(),
     SeveritiesParameter(),
     RepeatedMeasureParameter(),
-    ResourceFileParameter()
-), Scannable
+    ResourceFileParameter(),
+    AttachMediaParameter()
+), Scannable, ChartableData, DisplayValue