+++ app/src/main/java/com/fieldbook/tracker/ui/screens/traits/components/TraitOptionsSection.kt
@@ -29,6 +29,8 @@ import com.fieldbook.tracker.traits.formats.parameters.DecimalPlacesParameter
 import com.fieldbook.tracker.traits.formats.parameters.DefaultToggleParameter
 import com.fieldbook.tracker.traits.formats.parameters.InvalidValueParameter
 import com.fieldbook.tracker.traits.formats.parameters.MathSymbolsParameter
+import com.fieldbook.tracker.traits.formats.parameters.AttachMediaParameter
+import com.fieldbook.tracker.traits.formats.parameters.AllowOtherParameter
 import com.fieldbook.tracker.traits.formats.parameters.MultipleCategoriesParameter
 import com.fieldbook.tracker.traits.formats.parameters.Parameters
 import com.fieldbook.tracker.traits.formats.parameters.RepeatedMeasureParameter
@@ -146,7 +148,10 @@ fun TraitOptionsSection(
             formatDefinition?.let { definition ->
 
                 val displayableParams =
-                    definition.parameters.filter { it.parameter !in excludedParams }
+                    definition.parameters.filter { param ->
+                        param.parameter !in excludedParams &&
+                        !(param is AllowOtherParameter && isBrapiTrait)
+                    }
 
                 displayableParams.forEach { param ->
                     Chip(
@@ -209,6 +214,7 @@ private fun getParamIcon(param: BaseFormatParameter, trait: TraitObject): Int {
                 is InvalidValueParameter -> if (isEnabled) R.drawable.ic_outlier else R.drawable.ic_outlier_off
                 is MathSymbolsParameter -> if (isEnabled) R.drawable.ic_symbol else R.drawable.ic_symbol_off
                 is MultipleCategoriesParameter -> if (isEnabled) R.drawable.ic_multicat else R.drawable.ic_single_cat
+                is AllowOtherParameter -> if (isEnabled) R.drawable.ic_categorical_other else R.drawable.ic_categorical_other_off
                 is RepeatedMeasureParameter -> if (isEnabled) R.drawable.ic_repeated_measures else R.drawable.ic_repeated_measures_off
                 is SaveImageParameter -> if (isEnabled) R.drawable.ic_transfer else R.drawable.ic_transfer_off
                 else -> R.drawable.ic_tag_edit
@@ -220,6 +226,7 @@ private fun getParamIcon(param: BaseFormatParameter, trait: TraitObject): Int {
         is DecimalPlacesParameter -> R.drawable.ic_decimal
         is UnitParameter -> R.drawable.ic_tag_edit
         is ResourceFileParameter -> R.drawable.ic_tb_folder
+        is AttachMediaParameter -> R.drawable.multimedia
         else -> R.drawable.ic_tag_edit
     }
 }